requirements.txt
pytest
Pillow
ics
bcrypt
git+https://github.com/mattupstate/flask-security
flask-json
Babel==2.8.0
Click==7.0
decorator==4.4.1
dominate==2.4.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
flask-nav==0.6
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
Flask-WTF==0.14.3
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy
SQLAlchemy-Utils==0.36.1
validators==0.14.1
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
